+++ src/bika/lims/browser/__init__.py
@@ -26,7 +26,6 @@ from plone.memoize import view
 from Products.CMFCore.utils import getToolByName
 from Products.Five.browser import BrowserView as BaseBrowserView
 from senaite.core.api.dtime import to_localized_time as ulocalized_time
-from senaite.core.catalog import CONTACT_CATALOG
 from zope.cachedescriptors.property import Lazy as lazy_property
 from zope.i18n import translate
 
@@ -91,19 +90,9 @@ class BrowserView(BaseBrowserView):
     def checkPermission(self, perm, obj):
         return self.portal_membership.checkPermission(perm, obj)
 
-    # TODO: user_fullname is deprecated and will be removed in Bika LIMS 3.3.
-    # Use bika.utils.user_fullnameinstead
-    # I was having a problem trying to import the function from bika.lims.utils
-    # so i copied the code here.
+    @view.memoize
     def user_fullname(self, userid):
-        member = self.portal_membership.getMemberById(userid)
-        if member is None:
-            return userid
-        member_fullname = member.getProperty('fullname')
-        catalog = getToolByName(self, CONTACT_CATALOG)
-        c = catalog(portal_type='Contact', getUsername=userid)
-        contact_fullname = c[0].getObject().getFullname() if c else None
-        return contact_fullname or member_fullname or userid
+        return api.get_user_fullname(userid) or userid
 
     @view.memoize
     def user_email(self, userid):